renamed lockwidget files to uilocker
