+++ src/organizercore.h
@@ -14,7 +14,7 @@
 #include "usvfsconnector.h"
 #include "moshortcut.h"
 #include "processrunner.h"
-#include "lockwidget.h"
+#include "uilocker.h"
 #include <directoryentry.h>
 #include <imoinfo.h>
 #include <iplugindiagnose.h>
